requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
python-multipart==0.0.12
pydantic==2.10.3
pydantic-settings==2.6.1
supabase==2.10.0
httpx>=0.26,<0.28
PyJWT==2.10.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4 